marge TidyUpList.cs into DialogUI.cs
