--- XRObjects/Assets/XRObjects/Scripts/TidyUpList.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 7c1b4fbb1b14e4c468d1cd458286e41d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
